removed utility class for context in accessTokenProviderImp now it only stores token in repository and on the network callbacks (where we have context) store in shared preferences from the repository var as its the latest token.)
removed and deleted utility class for context in accessTokenProviderImp now it only stores token in repository and on the network callbacks (where we have context) store in shared preferences from the repository var as its the latest token.)

also used that variable in authentication fragment, and main gets it from shared and set it in repository too.

still pending to delete utility class i still have use for it in adapter, but i can get context from holder.view as Alan told me. doing that in the next commit.